Apply MVVM for profiles in SUI
