+++ src/cascadia/TerminalSettingsEditor/Microsoft.Terminal.Settings.Editor.vcxproj.filters
@@ -17,6 +17,7 @@
     <ClInclude Include="PreviewConnection.h" />
   </ItemGroup>
   <ItemGroup>
+    <Midl Include="ProfileViewModel.idl" />
     <Midl Include="EnumEntry.idl" />
     <Midl Include="Converters.idl">
       <Filter>Converters</Filter>
